% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorz.R
\name{\%..\%}
\alias{\%..\%}
\title{Binary operator to sample from a distribution.}
\usage{
pool \%..\% size
}
\arguments{
\item{pool}{A vector from where to sample}

\item{size}{The size of the sample}
}
\value{
A vector with sampled values
}
\description{
It's simply extracting without replacement.
}
\examples{
1:10 \%..\% 2
}
