% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorz.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Diff operator.}
\usage{
x \%<>\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The vector containing diff
}
\description{
Wraps the setdiff function.
}
\examples{
1:10 \%<>\% c(1, 3, 5)
}
