% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorz.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not in operator.}
\usage{
needle \%ni\% haystack
}
\arguments{
\item{needle}{A vector of things to check}

\item{haystack}{A vector of elements against which the check will be run}
}
\value{
A boolean vector
}
\description{
Basically, it is the opposite of %in%
}
\examples{
c(5, 42) \%ni\% 1:10
}
