% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorz.R
\name{\%||\%}
\alias{\%||\%}
\title{An OR operator for selecting one true value from two.}
\usage{
x \%||\% y
}
\arguments{
\item{y}{}
}
\description{
An OR operator for selecting one true value from two.
}
\examples{
NULL \%||\% 42
}
