% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorz.R
\name{\%||\%}
\alias{\%||\%}
\title{An OR operator for selecting one non NULL value from two.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The selected vector
}
\description{
An OR operator for selecting one non NULL value from two.
}
\examples{
x <- NULL \%||\% 42
}
