% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorz.R
\name{\%++\%}
\alias{\%++\%}
\title{Concat two vectors.}
\usage{
left \%++\% right
}
\arguments{
\item{left}{A vector}

\item{right}{A vector}
}
\value{
The pasted vector
}
\description{
Concat two vectors.
}
\examples{
"yo" \%++\% "da"
}
